Move `logo` role to a separate extension
