+++ orcestra_book/_config.yml
@@ -31,6 +31,7 @@ sphinx:
   local_extensions:
     apastyle: _ext/
     bracket_citation_style: _ext/
+    campaign_logos: _ext/
     flight_categories: _ext/
     front: _ext/
     reports: _ext/
